trying to fix 2 #261 add casting

+++ linkfd.c
@@ -3668,27 +3668,27 @@ int lfd_linker(void)
             add_json(js_buf, &js_cur, "rplp", "%d", info.i_rplp);
             add_json(js_buf, &js_cur, "frtt_Pus", "%d", info.frtt_us);
             add_json(js_buf, &js_cur, "frtt_appl", "%d", info.frtt_us_applied);
-            add_json(js_buf, &js_cur, "rtt2_lsn[1]", "%d", info.rtt2_lsn[1]);
+            add_json(js_buf, &js_cur, "rtt2_lsn[1]", "%u", (unsigned int)info.rtt2_lsn[1]);
             add_json(js_buf, &js_cur, "ertt", "%d", shm_conn_info->stats[info.process_num].exact_rtt);
-            add_json(js_buf, &js_cur, "buf_len", "%d", my_miss_packets_max);
-            add_json(js_buf, &js_cur, "buf_len_remote", "%d", miss_packets_max);
-            add_json(js_buf, &js_cur, "rsr", "%d", info.rsr);
+            add_json(js_buf, &js_cur, "buf_len", "%d", (int)my_miss_packets_max);
+            add_json(js_buf, &js_cur, "buf_len_remote", "%d", (int)miss_packets_max);
+            add_json(js_buf, &js_cur, "rsr", "%d", (int)info.rsr);
             add_json(js_buf, &js_cur, "rsr_top", "%d", rsr_top);
             add_json(js_buf, &js_cur, "sql", "%d", temp_sql_copy);
             add_json(js_buf, &js_cur, "pump_adj", "%d", pump_adj);
-            add_json(js_buf, &js_cur, "rtt_shift", "%d", rtt_shift);
-            add_json(js_buf, &js_cur, "W_cubic", "%d", info.send_q_limit_cubic);
-            add_json(js_buf, &js_cur, "send_q", "%d", send_q_eff);
+            add_json(js_buf, &js_cur, "rtt_shift", "%d", (int)rtt_shift);
+            add_json(js_buf, &js_cur, "W_cubic", "%u", (unsigned int)info.send_q_limit_cubic);
+            add_json(js_buf, &js_cur, "send_q", "%d", (int)send_q_eff);
             add_json(js_buf, &js_cur, "sqe_mean", "%d", send_q_eff_mean);
             //add_json(js_buf, &js_cur, "ACS", "%d", info.packet_recv_upload_avg);
-            add_json(js_buf, &js_cur, "ACS_ll", "%d", info.channel[1].ACS2);
+            add_json(js_buf, &js_cur, "ACS_ll", "%d", (int)info.channel[1].ACS2);
             add_json(js_buf, &js_cur, "ACS_rr", "%d", info.PCS2_recv * info.eff_len);
             add_json(js_buf, &js_cur, "ACS2", "%d", max_ACS2);
             add_json(js_buf, &js_cur, "PCS2", "%d", PCS * 2);
-            add_json(js_buf, &js_cur, "PCS_fast", "%d", info.channel[1].packet_download); // TMP REMOVE
+            add_json(js_buf, &js_cur, "PCS_fast", "%u", (unsigned int)info.channel[1].packet_download); // TMP REMOVE
             add_json(js_buf, &js_cur, "PCS_recv", "%d", info.PCS2_recv);
             //add_json(js_buf, &js_cur, "PCS_recvb", "%d", info.PCS2_recv * info.eff_len);
-            add_json(js_buf, &js_cur, "upload", "%d", shm_conn_info->stats[info.process_num].speed_chan_data[my_max_send_q_chan_num].up_current_speed);
+            add_json(js_buf, &js_cur, "upload", "%u", (unsigned int)shm_conn_info->stats[info.process_num].speed_chan_data[my_max_send_q_chan_num].up_current_speed);
             //add_json(js_buf, &js_cur, "dropping", "%d", (shm_conn_info->dropping || shm_conn_info->head_lossing));
             add_json(js_buf, &js_cur, "CLD", "%d", check_rtt_latency_drop()); // TODO: DUP? remove! (see CL below)
             //add_json(js_buf, &js_cur, "flush", "%d", shm_conn_info->tflush_counter);
